requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
pymysql
python-dotenv
openai
cryptography
Pillow
requests
geopy
earthengine-api
google-cloud-storage
gunicorn